fix(mobile): match the content-type case-insensitively on both bytes paths
`Headers` normalises header NAMES to lowercase but leaves VALUES alone, so
a `Content-Type: Text/Html` reached both exact-cased guards unchanged and
the login page was returned as certificate bytes — the regression test
without the fix hands back Uint8Array[60, 104, 116, 109], which is `<htm`.
CodeRabbit flagged client.ts; capacitorBinary.ts had the same defect on
the mobile path. The transport at :184 already lowercased.

Also splits the transport tests, which had grown to 485 lines against the
repo's 200-line convention: normalizeCapacitorBody moves beside the module
it actually tests, and fetchViaCapacitor splits by concern into request
shaping and response handling over a shared deps fixture. Same 216 tests
before and after — nothing added or dropped, only relocated.

And drops the explicit Content-Type from the plan's generateCert example,
which still taught the doubled-header bug 1eefff15 fixed; the spec claimed
the content-type "already matches what generateCert set by hand", which is
now false — it sets none.
